% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_misc.R
\name{check_numeric}
\alias{check_numeric}
\title{Check that a variable is numeric in a certain range}
\usage{
check_numeric(
  variable,
  min.value = 0,
  max.value = Inf,
  req.length = 1,
  check.whole = TRUE
)
}
\arguments{
\item{variable}{Value to check if numeric}

\item{min.value}{Minimum value for variable}

\item{max.value}{Maximum value for variable}

\item{req.length}{Required length for the variable.}

\item{check.whole}{Should the variabile be checked to see if it a whole number?}
}
\description{
Check that a variable is numeric in a certain range
}
