% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_misc.R
\name{format_p}
\alias{format_p}
\title{Format a p-value.}
\usage{
format_p(p, precision = 0.001)
}
\arguments{
\item{p}{P-value to be formatted}

\item{precision}{How many decimal places to be displayed in p-value}
}
\description{
Miscellaneous functions used in the routines for error checking and formatting output.
}
