% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_misc.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Check that argument is a whole number}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A value to be checked to see if it is whole number}

\item{tol}{Tolerance for checking if a whole number}
}
\description{
Check that argument is a whole number
}
