% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_calc_annual_stats.R
\name{stfl_calc_annual_stats}
\alias{stfl_calc_annual_stats}
\title{Compute the annualized statistics from daily data}
\usage{
stfl_calc_annual_stats(
  daily.data,
  Statistic = "Mean",
  water_year_start = 1,
  months_in_year = 12,
  roll_days = 1,
  min_roll_days = c(1, 3, 7, 30),
  roll_align = "right",
  ignore_missing = FALSE,
  STATION_NUMBER.var = "STATION_NUMBER",
  Date.var = "Date",
  Value.var = "Value",
  Parameter.var = "Parameter"
)
}
\arguments{
\item{daily.data}{Data frame (typically from the HYDAT data base) with the following fields
\itemize{

\item{STATION_NUMBER}{Station number, usually taken from the HYDAT database}
\item{Date}{Date of the reading in YYYY-MM-DD format}
\item{Parameter}{Type of measurement, typically either 'Flow'  or 'Level'}
\item{Value}{Value of the parameter for this date for this station. Numeric.}
}}

\item{Statistic}{A vector of statistics to be computed from the daily data.
The list of available statistics  is given gy \code{stfl_get_avail_stat()}.
A column in the returned data frame will have the statistic names.}

\item{water_year_start}{Start month for the water year.}

\item{months_in_year}{How many months, starting from the water_year_start should be
included in the water year. These two parameters allows you, for example, to select
June - Sept with \code{water_year_start=6} and \code{months_in_year=4}. If you specify
\code{water_year_start=10} and \code{months_in_year=4}, statistics will be computed
using October, November, December, and January (of the next year). The year in which the
water year starts is used as the label for the annualized value -- this differs from ther
\code{fasstr} package.}

\item{roll_days}{Numeric value of number of days to apply a rolling mean  (default 1) before statistics are computed.}

\item{min_roll_days}{Numeric value of number of days to apply a rolling minimum (and similar statistics) (default 1).
See \code{fasstr::calc_annual_stats()} for more details.}

\item{roll_align}{Character string identify the direction of the rolling mean from the specified date.
See \code{fasstr::calc_annual_stats()} for more details.}

\item{ignore_missing}{Should missing values be ignored when computing statistics. The default value of \code{FALSE}
implies that only complete water years (or portions selected using \code{water_year_start}) can be used in computing annualized statistics.}

\item{STATION_NUMBER.var}{Character string with the name of the variable containing the station number in the daily data and
in the output results..}

\item{Date.var}{Character string with the name of the variable containing the date.}

\item{Value.var}{Character string with the name of the variable containing the value to be analyzed.}

\item{Parameter.var}{Character string with the name of the variable containing the name of the Parameter in the daily data
and in the output results.}
}
\value{
Returns a data.frame with the following variables (sort order cannot be assumed)
\itemize{
\item{STATION_NUMBER}{The station number from the daily.data}
\item{Parameter}{The parameter from the daily data}
\item{Statistic}{The annualized statistics requested}
\item{Year}{The year the annualized statistic is computed}
\item{Value}{The value of the annualized statistic}
}
}
\description{
Compute the annualized statistics from daily data using the \code{fasstr} routines.
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- "08NM116"

daily <- hy_daily(station_number=station.id)
head(daily)

res <- stfl_calc_annual_stats(daily, Statistic=c("Mean","P90","Min"),
        water_year_start=1, months_in_year=2, ignore_missing=TRUE)
head(res1)
tail(res1)
}
}
