% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_extract_regional_predictions.R
\name{stfl_extract_regional_predictions}
\alias{stfl_extract_regional_predictions}
\title{Extract the regional predictions from a regional fit}
\usage{
stfl_extract_regional_predictions(
  fit,
  methods = c("LM", "MK", "LMrob"),
  transpose = FALSE
)
}
\arguments{
\item{fit}{Results from \code{stfl_trend_regional()} fit}

\item{methods}{Which methods should be used to estimate trend lines. Possible values are one or more of
\itemize{
\item{LM}{: Ordinary regression using the lm() function}
\item{MK}{: Mann-Kendall estimates using Sen estimator}
\item{LMrob}{: Robust linear regression fit using the lmrob() function from the robustbase package}

}}

\item{transpose}{If set to TRUE, then converts from wide-format to long-format.}
}
\value{
Returns a data.frame with the extracted predictions in a wide format
and in long format if \code{transpose=TRUE}.
}
\description{
Extract the regional predictions from a regional fit
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- c("08NM053","08NM116")
mydata <- stfl_get_annual_stat(station.id, Statistic=c("MEAN"))
mydata$REGION <- "AnyOldRegion"
mystation.weights <- data.frame(STATION_NUMBER=station.id, WEIGHT=c(2,1))
fit <- stfl_trend_regional(mydata, mystation.weights, log.Y=TRUE)
stfl_extract_regional_predictions(fit)
}
#'
}
