% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_extract_regional_trends.R
\name{stfl_extract_regional_trends}
\alias{stfl_extract_regional_trends}
\title{Extract the regional trend estimates}
\usage{
stfl_extract_regional_trends(
  fit,
  coefficient = c("slope", "intercept"),
  method = c("LM", "MK", "LMrob"),
  transpose = FALSE
)
}
\arguments{
\item{fit}{Results from \code{stfl_trend_regional()} fit.}

\item{coefficient}{Which coefficients to extract}

\item{method}{Which fitting method estimates to be extracted}

\item{transpose}{When \code{transpose=TRUE}, statistics are converted from long to wide format.}
}
\value{
Returns a data.frame with the extracted estimates, standard errors, p.values (and adjusted for serial autocorrelation)
}
\description{
Extract the regional trend estimates
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- c("08NM053","08NM116")
mydata <- stfl_get_annual_stat(station.id, Statistic=c("MEAN"))
mydata$REGION <- "AnyOldRegion"
mystation.weights <- data.frame(STATION_NUMBER=station.id, WEIGHT=c(2,1))
fit <- stfl_trend_regional(mydata, mystation.weights, log.Y=TRUE)
stfl_extract_regional_trends(fit)
}
#'
}
