% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_extract_station_predictions.R
\name{stfl_extract_station_predictions}
\alias{stfl_extract_station_predictions}
\title{Extract the station level predictions from a trend fit, both single-station-single-statistic and multi-station-multi-stat cases.}
\usage{
stfl_extract_station_predictions(
  fit,
  methods = c("LM", "MK", "LMrob"),
  transpose = FALSE
)
}
\arguments{
\item{fit}{Results from \code{stfl_trend_1stn_1stat()} or \code{stfl_trend_kstn_kstat()} fit.}

\item{methods}{Which methods should be used to estimate trend lines. Possible values are one or more of
\itemize{
\item{LM}{: Ordinary regression using the lm() function}
\item{MK}{: Mann-Kendall estimates using Sen estimator}
\item{LMrob}{: Robust linear regression fit using the lmrob() function from the robustbase package}

}}

\item{transpose}{If set to TRUE, then converts from wide-format to long-format.}
}
\value{
Returns a data.frame with the extracted predictions in a wide format
and in long format if \code{transpose=TRUE}.
}
\description{
Extract the station level predictions from a trend fit, both single-station-single-statistic and multi-station-multi-stat cases.
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- "08NM116"
data <- stfl_get_annual_stat(station.id, Statistic=c("MEAN"))
res <- stfl_trend_1stn_1stat(data, log.Y=TRUE)
stfl_extract_station_predictions(res)
stfl_extract_station_predictions(res, transpose=TRUE)
stfl_extract_station_predictions(res, transpose=TRUE, methods="LM")
}
#'
}
