% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_extract_station_trends.R
\name{stfl_extract_station_trends}
\alias{stfl_extract_station_trends}
\title{Extract the station level estimates from a trend fit, both single-station-single-statistic and multi-station-multi-stat cases.}
\usage{
stfl_extract_station_trends(
  fit,
  coefficient = c("slope", "intercept"),
  methods = c("LM", "MK", "LMrob"),
  transpose = FALSE
)
}
\arguments{
\item{fit}{Results from \code{stfl_trend_1stn_1stat()} or \code{stfl_trend_kstn_kstat()} fit.}

\item{coefficient}{Which coefficients to extract}

\item{methods}{Which methods should be used to estimate trend lines. Possible values are one or more of
\itemize{
\item{LM}{: Ordinary regression using the lm() function}
\item{MK}{: Mann-Kendall estimates using Sen estimator}
\item{LMrob}{: Robust linear regression fit using the lmrob() function from the robustbase package}

}}

\item{transpose}{If set to TRUE, then converts from long-format to wide-format with the statistics from the methods all on one line.}
}
\value{
Returns a data.frame with the extracted estimates, standard errors, p.values (and adjusted for serial autocorrelation) in a long format
and in wide format if \code{transpose=TRUE}.
}
\description{
Extract the station level estimates from a trend fit, both single-station-single-statistic and multi-station-multi-stat cases.
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- "08NM116"
data <- stfl_get_annual_stat(station.id, Statistic=c("MEAN"))
res <- stfl_trend_1stn_1stat(data, log.Y=TRUE)
estimates <- stfl_extract_station_trends(res)
}
#'
}
