% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_options.R
\name{stfl_options}
\alias{stfl_options}
\title{Options for this package.}
\usage{
stfl_options()
}
\value{
Returns a list with options
}
\description{
Options for this package. Use the settings::options_manager() to create an override function
to change the options here
}
