% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_plot_add_regional_fit.R
\name{stfl_plot_add_regional_fit}
\alias{stfl_plot_add_regional_fit}
\title{Add the regional fit to an existing trend plot}
\usage{
stfl_plot_add_regional_fit(
  plot,
  regional.fit,
  plot.fit = c("LMrob", "LM", "MK")[1]
)
}
\arguments{
\item{plot}{An existing plotly objects to which the regional fit will be added}

\item{regional.fit}{A list object containing the results of the region fit. See
the help page for \code{stfl_trend_regional()} for details on this object}

\item{plot.fit}{Which trend lines should be added to the plots. Possible values are one or more of
\itemize{
\item{LM}{: Ordinary regression using the lm() function}
\item{MK}{: Mann-Kendall estimates using Sen estimator}
\item{LMrob}{: Robust linear regression fit using the lmrob() function from the robustbase package}

}
If you try and plot a method that was not computed, a warning message is printed and the request is ignored.}
}
\value{
Plotly object is returned with the regional fit(s) added to the plot.
}
\description{
Add the regional fit to an existing trend plot
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- c("08NM053","08NM116")
data <- stfl_get_annual_stat(station.id, Statistic=c("MEAN"))
temp <- stfl_trend_kstn_kstat(data, log.Y=TRUE)
myplot <- stfl_plot_trends(temp, plot.fit=c("MK","LM","LMrob"))
myplot
data$REGION <- "AnyOldRegion"
mystation.weights <- data.frame(STATION_NUMBER=station.id, WEIGHT=c(2,1))
regional.fit <- stfl_trend_regional(data, mystation.weights, log.Y=TRUE)
stfl_plot_add_regional_fit(myplot,  regional.fit[[1]])
}
}
