% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_plot_trends.R
\name{stfl_plot_trends}
\alias{stfl_plot_trends}
\title{Plot the results from station level fits to various statistics and various stations}
\usage{
stfl_plot_trends(
  fit.res,
  plot.fit = c("LMrob", "LM", "MK")[1],
  connect.points = TRUE,
  force.Y.scale.log = FALSE,
  plotly = TRUE
)
}
\arguments{
\item{fit.res}{Fit result created by \code{tfl_trend_kstn_kstat()} which is a list and each element of the list has
the \code{estimates} data.frame and the \code{data.aug} data.frame with original data with the predicted values from a
LM, MK, and LMrob fit.}

\item{plot.fit}{Which trend lines should be added to the plots. Possible values are one or more of
\itemize{
\item{LM}{: Ordinary regression using the lm() function}
\item{MK}{: Mann-Kendall estimates using Sen estimator}
\item{LMrob}{: Robust linear regression fit using the lmrob() function from the robustbase package}

}
If you try and plot a method that was not computed, a warning message is printed and the request is ignored.}

\item{connect.points}{Should the annualized values be connected with a line.}

\item{force.Y.scale.log}{Should the Y axis be plotted on the log-scale regardless if the
fits are done on the log scale. Note that if the original fit was NOT done on the log(Y) scale,
then setting \code{force.Y.scale.log=TRUE} will NOT change the fit - only the ysalce}

\item{plotly}{Should a plot_ly() graph be returned (default), or a ggplot graph}
}
\value{
plot_ly graph of the results
}
\description{
Plot the results from station level fits to various statistics and various stations
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- "08NM116"
data <- stfl_get_annual_stat(station.id, Statistic=c("MEAN","P10","P90"))
temp <- stfl_trend_kstn_kstat(data, log.Y=TRUE)
stfl_plot_trends(temp, plot.fit=c("MK","LM","LMrob"))
stfl_plot_trends(temp, plot.fit=c("LMrob"), plotly=TRUE)
}
}
