% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfl_trend_regional.R
\name{stfl_trend_regional}
\alias{stfl_trend_regional}
\title{Regional analysis on trends
Computes a regional analysis using least squares (LM), Mann-Kendall (MK), or robust least squares (LMrob)
with HAC corrections for autocorrelation.
This computes the individual slopes for each station within each region, and then weights the
individual slopes using the weights.}
\usage{
stfl_trend_regional(
  data,
  station.weights,
  log.Y = FALSE,
  REGION.var = "REGION",
  STATION_NUMBER.var = "STATION_NUMBER",
  WEIGHT.var = "WEIGHT",
  Year.var = "Year",
  Statistic.var = "Statistic",
  Value.var = "Value",
  Parameter.var = "Parameter",
  offset = NULL,
  nsim.bbmks = 2000,
  trend.min.n = stfl_options()$trend.min.n
)
}
\arguments{
\item{data}{\itemize{
\item data frame with REGION STATION NUMBER, Parameter, Statistic, Year (X) and Value (Y) variables
}}

\item{station.weights}{\itemize{
\item data frame with STATION_NUMBER, and WEIGHT
}}

\item{log.Y}{Should the analysis be done on the log(Y+offset) scale where log() referes to natural logarithms.
If log(y) makes no sense (e.g., if statistic is DoY), the log() is ignored and set to FALSE automatically.
Refer to the \emph{offset} argument for more details.

In most cases, the analysis on the \emph{log(Y)} scale is preferred because of the simple interpretation of the trend.
For example, an slope of .02 on the \emph{log(Y)} corresponds to a 2\% increase/year regardless of the underlying units.

In some cases, a \emph{log(Y)} transform makes no sense. For example, the Day_of_Year (DoY) when a minimum occurs is arbitrary
and depends on when the start of the year begins (e.g. a water year starts on 1 October).

The \emph{stfl_get_avail_stat()} function has a list of available statistics and if a \emph{log(Y)} transform is allowable.}

\item{REGION.var}{Character string with the name of the variable containing the REGION in the data and
in the output results..}

\item{STATION_NUMBER.var}{Character string with the name of the variable containing the station number in the daily data and
in the output results..}

\item{WEIGHT.var}{Character string with the name of the variable containing the relative weight to be used
when finding thre regional estimates. These are normalized to sum to 1 internally.}

\item{Year.var}{Name of the Year variable (usually the year the data was collected)}

\item{Statistic.var}{Character string with the name of the variable containing the statistic that was computed}

\item{Value.var}{Character string with the name of the variable containing the value to be analyzed.}

\item{Parameter.var}{Character string with the name of the variable containing the name of the Parameter in the daily data
and in the output results.}

\item{offset}{What is the value of the offset used when taking log(Y+offset).
If unspecified, then the offset is 1/2 of the smallest positive values.

The \emph{offset} is needed to avoid taking log(0).}

\item{nsim.bbmks}{\itemize{
\item Number of bootstrapped simulations for Mann-Kendall block boot strap
}}

\item{trend.min.n}{\itemize{
\item Minimum number of data points for trend computation
}}
}
\value{
List with a list for every station containing the station trends  and additional entries for region with the following elements
\itemize{
\item{data.aug}{Data frame augmented with the LM, MK, LMrob station trend predictions}
\item{estimates}{Data frame with estimates and standard errors and the following columns for each station
x, y, log.Y, estimate (intercept or Y variable), std error (ignoring autocorrelation),
p.value (ignoring autocorrelation), std.error adjusted for HAC}
\item{regional.estimate}{Data frame with similar variables but with a weighted average of
trends trends from the stations}
\item{regional.pred}{Data frame augmented with the LM, MK, LMrob regional trend predictions}
}
}
\description{
Regional analysis on trends
Computes a regional analysis using least squares (LM), Mann-Kendall (MK), or robust least squares (LMrob)
with HAC corrections for autocorrelation.
This computes the individual slopes for each station within each region, and then weights the
individual slopes using the weights.
}
\examples{
\dontrun{
# needs the HYDAT package installed
station.id <- c("08NM053","08NM116")
mydata <- stfl_get_annual_stat(station.id, Statistic=c("MEAN"))
mydata$REGION <- "AnyOldRegion"
mystation.weights <- data.frame(STATION_NUMBER=station.id, WEIGHT=c(2,1))
stfl_trend_regional(mydata, mystation.weights, log.Y=TRUE)
}

}
